refactor: added the necessary structural components, refactored app
