+++ src/app/auth/auth.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@NgModule({
+    declarations: [],
+    imports: [
+        CommonModule
+    ]
+})
+export class AuthModule { }
