+++ src/app/core/components/filtering-criteria-block/filtering-criteria-block.component.spec.ts
