+++ src/app/core/components/filtering-criteria-block/filtering-criteria-block.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { FilterService } from 'src/app/services/filter/filter.service';
-import { SortService } from 'src/app/services/sort/sort.service';
+import { FilterService } from 'src/app/youtube/services/filter/filter.service';
+import { SortService } from 'src/app/youtube/services/sort/sort.service';
 
 @Component({
     selector: 'app-filtering-criteria-block',
