+++ src/app/core/components/header-search-block/header-search-block.component.html
@@ -7,10 +7,7 @@
         id="searchQuery"
         (keyup)="onEnterKey($event)"
     />
-    <app-custom-button
-        additionalClasses="search__input-button"
-        type="submit"
-    >
+    <app-custom-button additionalClasses="search__input-button" type="submit">
         search
     </app-custom-button>
 </form>
