+++ src/app/core/components/header-search-block/header-search-block.component.scss
