+++ src/app/core/components/logo/logo.component.scss
