+++ src/app/core/components/logo/logo.component.spec.ts
