+++ src/app/youtube/components/metrics/metrics.component.scss
