+++ src/app/youtube/components/metrics/metrics.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { SearchItem } from 'src/app/models/search-item.model';
+import { SearchItem } from 'src/app/youtube/models/search-item.model';
 
 interface Metrics {
     icon: string;
