+++ src/app/youtube/components/search/search-item/search-item.component.scss
