+++ src/app/youtube/components/search/search-results/search-results.component.scss
