+++ src/app/youtube/directives/colorByDate/color-by-date.directive.ts
